Refactor following peer review. Move XML formatting for -webapp-kogito-testing to new JS file.
