+++ kie-wb-common-dmn/kie-wb-common-dmn-webapp-kogito-testing/src/main/webapp/index.html
@@ -29,6 +29,7 @@
     <script type="text/javascript" src="model/DMN12.js"></script>
     <script type="text/javascript" src="model/KIE.js"></script>
     <script type="text/javascript" src="model/MainJs.js"></script>
+    <script type="text/javascript" src="model/FormatterJs.js"></script>
 
 </head>
 <body>
